Change the return type of counting methods in various DAOs from int to long
This is required for job repository/explorer implementations
that use the Long type to count records in the database.

Resolves #4227

+++ spring-batch-core/src/main/java/org/springframework/batch/core/explore/JobExplorer.java
@@ -153,6 +153,6 @@ public interface JobExplorer {
 	 * @throws NoSuchJobException thrown when there is no {@link JobInstance} for the
 	 * jobName specified.
 	 */
-	int getJobInstanceCount(@Nullable String jobName) throws NoSuchJobException;
+	long getJobInstanceCount(@Nullable String jobName) throws NoSuchJobException;
 
 }
